+++ src/http-server/client/response.js
@@ -1,7 +1,7 @@
 // @ts-check
 
 export default class VelociousHttpServerClientResponse {
-  /** @type {string | null} */
+  /** @type {string | Uint8Array | null} */
   body = null
 
   /** @type {Record<string, string[]>} */
@@ -40,7 +40,7 @@ export default class VelociousHttpServerClientResponse {
   }
 
   /**
-   * @returns {string | null} - The body.
+   * @returns {string | Uint8Array | null} - The body.
    */
   getBody() {
     if (this.body !== undefined) {
@@ -65,7 +65,7 @@ export default class VelociousHttpServerClientResponse {
   }
 
   /**
-   * @param {string} value - Value to use.
+   * @param {string | Uint8Array} value - Value to use.
    * @returns {void} - No return value.
    */
   setBody(value) {